python/pyproject.toml
[project]
name = "python-realtime"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "langchain-core",
    "langchain",
    "python-dotenv"
]

[project.optional-dependencies]
elevenlabs = ["elevenlabs"]
anthropic = ["langchain-anthropic"]
